[AUDIO] Support new audio devices enum values introduced in JB MR1 code drop (part 2)
BZ: 71815

The check whether a route or a strategy is applicable is not correct.
The direction bit of an input device is set to 1.
So the AND check with a mask will always result to 'true'.
The solution is to remove the bit direction before testing.
But alsa_mfld_cdk still needs the direction info. So this is
added as argument in alsa_handle_t.

Change-Id: Iafe3a7692d312982c9359f31897a803ae6336078
Signed-off-by: pafonsoX <philippex.afonso@intel.com>
Reviewed-on: http://android.intel.com:8080/79263
Reviewed-by: Centelles, Sylvain <sylvain.centelles@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>