[AUDIO] Rebase IA controlled SSP rte mgr on JB MR1
BZ: 73258

Rebase the route manager for CTP and derivated on JB MR1.
It also updates the LPE centric route manageR.
The main changes is that input devices uses the sign bit.
All the strategy of the route manager is base on bit field,
so the route manager needs to remove the MSB.

Change-Id: I989670cc7f09c16b6303a8390e1abb2d335b24a9
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/81012
Reviewed-by: Barthes, FabienX <fabienx.barthes@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>