[AUDIO] Recent skype application: no audio
BZ: 75043

No audio can be heard with last skype application when using
new route manager.
The latest skype app makes use of android mode, but it does not
select the incommunication input source. As the new route manager
has a strong restriction on the VoIP path, it does not allow
the UL to be routed.
This patch add flexibility on the input source, the integrity of
the VoIP route is completely kept by checking upon the android
mode for BOTH playback and capture.

Change-Id: I5a9df9cc14918198560588441fca110a0cedfac0
Origin-Change-Id: I1c6e473cdc11620970855b60cf2c454c05de8270
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/83039
Reviewed-by: Barthes, FabienX <fabienx.barthes@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>