[AUDIO] Stream Offload Management with new route manager
BZ: 75313

Offload use cases were broken with introduction of new route
manager.
This patch changes the way policy and primary HAL (so Route manager)
communicate. It now use a setParameter with a StreamFlag key
when a compressed stream starts and stops.
It also still send the accessory change on the primary HAL when an
offload use case is on going.
Media and Compressed are now exclusive.

Change-Id: Ic912430789849de968261108d03d6c7d2cf703d4
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/83481
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>