[Audio] New modem audio manager design
BZ: 79591

The new modem audio manager design provides a new interface on which
the route manager should connect to interact with.

This ModemAudioManagerInterface abstracts the modem manager
implementation which can be different from one modem to another.

This interface has to be retrieved by the route manager from the new
libmamgr.so library.

Setting itself as an observer of this interface, the route manager
will be notified on modem events including:
- Modem state changes (up, down)
- Modem audio availability changes
- Modem audio band change (narrow, wide)

Change-Id: I4c8628edc05b08a3ec93b6768af440f1380ec843
Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>
Reviewed-on: http://android.intel.com:8080/86206
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>