[AUDIO] Lock contention between playback and capture
BZ: 77322

Playback is in background, launch a capture: playback is not smooth.
Each read/write does a setStandby(false) on the route manager.
It takes the routing lock in write mode to ensure whether the stream
is started or not, blocking any concurrent read/write.
This patch make use of the routing lock in read mode to access
the member variable of the stream.

Change-Id: Ic5281f4baf932aa9e5f1bdf47fc9200d097d86a5
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/86306
Reviewed-by: Di Folco, Neil <neil.di.folco@intel.com>
Reviewed-by: Quintero, Jorge <jorge.quintero@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>