[AUDIO][WORKAROUND] Recovery to Audio loss in Touch sounds
BZ: 80798

Issue: Codec requires some settling time and an addtional
delay of 20ms for SRC in codec. This is a limitation in the
codec, so the short audio tones like touch tones are not
heard (lost) because of the delay in power up.

Fix: [WORKAROUND] A virtual device is opened always when the
display is ON, this will help in keeping the audio codec
setup quick
Only the 20ms silence is written at the start
of the first write call.

Change-Id: I10cab014d71e162957081a4217a16102575aec37
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/87374
Reviewed-by: Di Folco, Neil <neil.di.folco@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Centelles, Sylvain <sylvain.centelles@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>