[AUDIO][REVERTME] Force BT always to enabled on VB platform
BZ: 82673

In order to enable the BT call from PhoneApp UI, Bluetooth
needs to be forced to enabled in the audio route manager, as
the key parameter from BT service indicating that BT is ON or
OFF is not received anymore by the audio route manager.
This patch must be reverted once the complete Bluetooth stack
will be integrated on JB (Android 4.2.1), and the BT
parameter key will correctly inform the audio route manager.

Change-Id: I7a55d1b8d7c49fd0fd1b69844bd53fc57914c2c3
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>
Reviewed-on: http://android.intel.com:8080/88641
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Reviewed-by: Gaffie, Francois <francois.gaffie@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>