[PORT FROM JBMR1][AUDIO] Avoid extra reconfigure on screen event
BZ: 87147

When the screen switches to on or to off, and while a playback
is on going (whatever compressed or not), audio cut can be heard.
Some mixers are switched on then off during the reconsider routing.
Need to filter the condition of needreconfiguration of HwCodecMedia
route and change the rules of mute / unmute configuration of these
controls mixer.

Change-Id: I15be0396389544baf4156398cc70cbbba0665621
Origin-Change-Id: I1a8ba0da2988cf7d8cf1d2478b61a6cf5bd4aa8b
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/95124
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>