[AUDIO] Avoid extra reconfigure on screen event
BZ: 87147

When the screen switches to on or to off, and while a playback
is on going (whatever compressed or not), audio cut can be heard.
Some mixers are switched on then off during the reconsider routing.
Need to filter the condition of needreconfiguration of HwCodecMedia
route and change the rules of mute / unmute configuration of these
controls mixer.

Change-Id: I1a8ba0da2988cf7d8cf1d2478b61a6cf5bd4aa8b
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/92507
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Di Folco, Neil <neil.di.folco@intel.com>
Reviewed-by: Gonzalve, Sebastien <sebastien.gonzalve@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>