[PORT FROM JBMR1][AUDIO] Use derivated class of Media Audio Route
BZ: 88340

Loss touch sound WA was broken when the patch of followup
of the review was merge. The Media route was not using
the appropriate derivated class, and as a result, no silence
were appended when starting to play.
This patch restores the instantiation of the derivated class
for media stream route.

Change-Id: Ie53ed211cbcfb8270a7573445f206d230054ddd4
Origin-Change-Id: Id7850f4b37776ad71a908ceb3a456576e140f2fe
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>