[AUDIO] Use derivated class of Media Audio Route
BZ: 88340

Loss touch sound WA was broken when the patch of followup
of the review was merge. The Media route was not using
the appropriate derivated class, and as a result, no silence
were appended when starting to play.
This patch restores the instantiation of the derivated class
for media stream route.

Change-Id: Id7850f4b37776ad71a908ceb3a456576e140f2fe
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/93010
Reviewed-by: Di Folco, Neil <neil.di.folco@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Gonzalve, Sebastien <sebastien.gonzalve@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>