[AUDIO][PFW] keep SPEECH_RX source enabled in modem when swapping from CSV to VOIP
BZ: 79826

As per IMC recommendation, during a CSV => VoIP swap, the SPEECH_RX has to
be kept ON due to fact that call is not disconnected. Keeping the source
active is required so that VOCODER will not be reset.
This patch creates a new criterion based on modem audio state that makes
distinction between 'end-of-csv-call' and 'swap-csv-voip' situations.

Change-Id: I69fdc10c310b06d2bacd38d16297559a9495b024
Signed-off-by: pdiopX <piotrx.diop@intel.com>
Reviewed-on: http://android.intel.com:8080/99739
Reviewed-by: Di Folco, Neil <neil.di.folco@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>