[Audio] Enable BT HFP 1.6 wide band support with B+
BZ: 95869

The BT HFP 1.6 is not supported yet.

This patch introduces a band parameter for the bluetooth accessory.
The HFP, from version 1.6, introduces a band negociation with the BT
device. While BT band was always narrow, it now can be narrow or wide
according to the BT device capability.

The patches updates the Audio Route Manager to exposes a new platform
state to reflect the BT negociated audio band.

Moreover, it updates the Audio Route Manager to retrieve the negociated
BT audio band from a setParameter.

Since HFP 1.6 is not yet handled natively by Android, the parameter
which provides the BT audio band is not standard. This parameter is
defined and managed by the B+ stack.

Change-Id: I47b044f5fbf4e45e70b120f356071438d21693b0
Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>
Reviewed-on: http://android.intel.com:8080/99936
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>