[Audio] 2 seconds noise after answering MT Skype call
BZ: 98370

2 seconds noise can be heard from DUT speaker after answering MT voip call via
Skype. This is due to a setStrategyMute putting 2 seconds of sleep when
the phone goes from ringtone to in communication mode.
During this "sleep" time, the output device is still IHF. The earpiece is
routed afterwards.

The solution is to force the device to earpiece as soon as the phone
state is in communication. This can be used  for CSV call.

Change-Id: I704415eded4cc89e30188ddd35b7fafa9cd0870b
Signed-off-by: pafonsoX <philippex.afonso@intel.com>
Reviewed-on: http://android.intel.com:8080/101546
Reviewed-by: Di Folco, Neil <neil.di.folco@intel.com>
Reviewed-by: Becker, VincentX <vincentx.becker@intel.com>
Reviewed-by: Cahuzac, ThomasX <thomasx.cahuzac@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>