[AUDIO] Add new tinyalsa prepare API in Audio HAL
BZ: 101353

Alloc stream is made upon prepare ioctl. It is called on the first pcm_write, so
after all routing commands. Similar issue for the stop, freeding the stream
is done before the routing command.
MRLD LPE does not support this sequence.

This patch restablishes the MRFLD sequence agreed in the FAS/SAS
by making use of the prepare API called when opening the device alsa.
It allocates the stream before sending any AUD routing commands.

Change-Id: I649fbdb3b8f4acb2eaf1842327469d0ab38ae0ff
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/101776
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Gonzalve, Sebastien <sebastien.gonzalve@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>