[AUDIO][REVERTME] Open conditionnally the PCM route in config stage
BZ: 96542

Routes in the LPE centric audio arch shall be enabled in a locked context.

In a locked context, the audio shall still be opened. Until direct
output will be implemented, this patch allows the routes to be
opened at configuration stage. And at enable stage, the routes
shall not be opened.

With use of a define, enable the routes at the configuration stage.

Change-Id: Iad6d1f2e7bf8cf3a4e561ea971138018e2466113
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>
Reviewed-on: http://android.intel.com:8080/102347
Reviewed-by: Gaffie, Francois <francois.gaffie@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Cahuzac, ThomasX <thomasx.cahuzac@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>