[PORT FROM R42][AUDIO] Configure the volume keys GPIOs when in CSV call.
BZ: 97726

Today, the user cannot adjust the volume during a CSV call when the DUT is in
S3 state. This is due to the fact the volume key GPIOs are not configured
as S3 state wakeup sources.

This fix configures the volume keys GPIOs as S3 state wakeup sources only when
in CSV call. In any other cases, the volume key GPIOs are not configured as S3
state wakeup sources in order to avoid extra power consumption.

Change-Id: Ib46ba089ecec2325f133be1271ba6aa90a555626
Origin-Change-Id: I066b8d940390a29895bfbc323dac0e1310236ffe
Signed-off-by: pafonsoX <philippex.afonso@intel.com>
Reviewed-on: http://android.intel.com:8080/105382
Reviewed-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>