[AUDIO] Keypress tone shall not be heard while MO/MT call in silent mode
BZ: 96985

In ALSA policy, the volume computation in call does not check the value of
the SYSTEM stream volume index and return the max volume even in silent
mode.

This fix compares the SYSTEM stream volume index value to the min SYSTEM
stream volume index value. If both values are equal, the SYSTEM stream
volume won't be updated.

Change-Id: Ia1bb3c57e84520a516b8a605147d8150e5ac1bdc
Signed-off-by: Fabien Rodriguez <fabienx.rodriguez@intel.com>
Reviewed-on: http://android.intel.com:8080/106322
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: Fomitchev, AndreiX <andreix.fomitchev@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>