[AUDIO] Avoid hearing alarm beep sound and alarm sound together in call
BZ: 102414

When setting an alarm during a call, alarm beep sound and alarm sound can
be heard together, because of a bad volume computation of ALARM and
VOICE_CALL streams in ALSA policy.

When setting an alarm during a call, only alarm beep sound shall be
heard. Alarm beep sound is using the VOICE_CALL stream and alarm sound is
using the ALARM stream. To avoid playing both sounds together, the volume
computation in ALSA policy shall return the max volume on VOICE_CALL
stream instead of ALARM stream (when alarm sound is played, its index is
null so its volume will be automatically muted).

Note that VOIP call is also using the VOICE_CALL stream but VOIP volume
won't be impacted by this change thanks to the following patch :
http://android.intel.com:8080/#/c/93376/

Change-Id: I8f77096fcd4da627c0a294009eafb0b917f5b840
Signed-off-by: Fabien Rodriguez <fabienx.rodriguez@intel.com>
Reviewed-on: http://android.intel.com:8080/106321
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Barthes, FabienX <fabienx.barthes@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>