[PDK MR2 Temp] Do not rely on the BT state to elect the BT route
BZ: 117199

Instead, rely solely on the policy information regarding the device selection.

We used to use the 'bluetooth_enabled' parameter but this information was lost
in R4 and reintroduced in
frameworks/base/services/java/com/android/server/BluetoothManagerService.java
by http://android.intel.com:8080/#/c/91943/
However, this file is not delivered as source in the PDK.

This patch may be merged to mainline but this needs testing.

Change-Id: I714d5a36ce7c6b4e4fb930849c6899f04df7f10b
Signed-off-by: David Wagner <david.wagner@intel.com>