[AUDIO] RouteManager - Add a dependency of Voice Route to the Bt route
BZ: 107515

When performing a VoIP call over a Bluetooth device, the following
output/input route list should be applied:
Voice | BtIA
In order to achieve that, the "Voice" route has to be declared
as a slave route of the BT route.

This commit adds "Voice" to the slave route list of "Bt" route.

Change-Id: Ib03aabcfa4506d5d7257900a71b7cd8697712fc1
Signed-off-by: GabrielX Malgouyard <gabrielx.malgouyard@intel.com>
Reviewed-on: http://android.intel.com:8080/107315
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Becker, VincentX <vincentx.becker@intel.com>
Reviewed-by: Gaffie, Francois <francois.gaffie@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>