[WM8958] Property ro.AudioComms.PFW.ConfPath should not be read only
BZ: 103991

The Android property ro.AudioComms.PFW.ConfPath is used to give the
pathname of the main PFW configuration file. It should be possible to
set the correct path depending on runtime parameters, thus this
property should not be readonly.

Changed the name of the property to AudioComms.PFW.ConfPath

Change-Id: I080db1f1b4c08f2301376126277e4cf26bf268d1
Signed-off-by: Frederic Boisnard <fredericx.boisnard@intel.com>
Reviewed-on: http://android.intel.com:8080/108697
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>