[AUDIO] Avoid extra reconfigure on screen event
BZ: 110978

When the screen switches to on or to off, and while a playback is on going
(whatever compressed or not, in call or not), audio cut can be heard.
Some mixers are switched on then off during the reconsider routing.

Then we filter the condition of needreconfiguration of HwCodecMedia
route for Out-of-call UCs, and HwCodecCSV route for In-call UCs.

Change-Id: Iaae91d0b6b6c8698dc4a7bd02e266f2cc2d7d87b
Signed-off-by: ptonnetx <pierre-louisx.tonnet@intel.com>
Reviewed-on: http://android.intel.com:8080/110131
Reviewed-by: Gaffie, Francois <francois.gaffie@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>