[AUDIO] Enable HwCodec routes reconfiguration
BZ: 114566

Currently, HwCodec audio routes are not reconfigured when
touch sounds are enabled. As a result, when touch tones are
enabled and switching to IHF during call, the speech remains
on earpiece. A reconfiguration of the HwCodec routes is
needed, in all cases the route changes.

Remove the call to the method needReconfiguration with the
parameter 'false' in HwCodec routes definitions.

Change-Id: Id5412a5ae1aaa1c3187e3b1489918f237646f469
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>
Reviewed-on: http://android.intel.com:8080/112142
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>