[AUDIO] Add AudioPlatformHardware file for byt-t-ffrd8 variant
BZ: 117369

Today, in the byt-t-ffrd8 variant there is not AudioPlatformHardware file.
It compiles because the REF_PRODUCT_NAME is equal to baylake for the byt-t-ffrd8 BYT variant.

In the first step , this file will be created as symbolic link of AudioPlatformHardware_baylake.cpp.
Once both files (AudioPlatformHardware_baylake.cpp and AudioPlatformHardware_byt-t-ffrd8.cpp) diverge,
one file will be created and the symbolic link destroyed.

Change-Id: I90d8e2aec8ee72d75f0359b2ab92db3ce2aae2da
Signed-off-by: Eduardo Mendi <eduardox.mendi@intel.com>
Reviewed-on: http://android.intel.com:8080/114270
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Di Folco, Neil <neil.di.folco@intel.com>
Reviewed-by: Denneulin, Guillaume <guillaume.denneulin@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Reviewed-by: Molero Rodenas, BeatrizX <beatrizx.molero.rodenas@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: Tinelli, Vincent <vincent.tinelli@intel.com>
Tested-by: Tinelli, Vincent <vincent.tinelli@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>