[AUDIO] Make distinction between stopped and not-yet-started input
BZ: 115755

There was a race condition when a client calls getInput(), but is preempted
by a second client getInput() call for the same source. In that case, after
the release of the latter, the input is given back to the first client if he
is still alive (see patch I57b4b27151475d449d503d8e12d7d57159d04d77).
In this situation the startInput() call from the first client will fail as
the policy assumes that it has already started and is active.

This fix implements a new inputDescriptor parameter, mHasStarted to distinguish
an input that has been stopped (and therefore already called startInput()) from
an input that did not yet start.

Change-Id: I3de3f673757317cc26b8b8055543bdafaa12d4ff
Signed-off-by: pdiopX <piotrx.diop@intel.com>
Reviewed-on: http://android.intel.com:8080/116552
Reviewed-by: Di Folco, Neil <neil.di.folco@intel.com>
Reviewed-by: Gonzalve, Sebastien <sebastien.gonzalve@intel.com>
Reviewed-by: Gaffie, Francois <francois.gaffie@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>