[Audio] Disable the BtComm route when IHF+SCO are the output devices
BZ: 121257

Now the RouteManager set as applicable only the first current route.
But if there is more than one route applicable, this can lead to the wrong one.
For instance, if the BtComm is set before the HwCodecMedia in the
AudioPlatformHardware then the output can be BT SCO instead of IHF.
This can happen if the user takes a picture during a VOIP call on BT SCO.

The solution is to check if BT SCO is the only possible device before setting it
as applicable.

Change-Id: Ic3fd1e0da725b56f9e5429e4443122492cbb9e39
Signed-off-by: pafonsoX <philippex.afonso@intel.com>
Reviewed-on: http://android.intel.com:8080/118834
Reviewed-by: Gaffie, Francois <francois.gaffie@intel.com>
Reviewed-by: Di Folco, Neil <neil.di.folco@intel.com>
Reviewed-by: Gonzalve, Sebastien <sebastien.gonzalve@intel.com>
Reviewed-by: Quintero, Jorge <jorge.quintero@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>