[AUDIO] Dynamic audio dump library implementation
BZ: 127147

Create a new library for dynamic audio dumps in the
route manager. This will replace the former activation
of audio dumps in the audio HAL by a compilation macro.

The implementation is part of a library linked to the
route manager. The library's dump interface is called
from probe points activated dynamically by the user
with Android properties.

It is planned to be integrated in the Audio Toolbox.

Change-Id: Ic9a272636b824bd86c1f0d36b1a848b94a84f69e
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>
Reviewed-on: http://android.intel.com:8080/122428
Reviewed-by: Gonzalve, Sebastien <sebastien.gonzalve@intel.com>
Reviewed-by: Quintero, Jorge <jorge.quintero@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: Gaffie, Francois <francois.gaffie@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>