[AUDIO] Generate tombstone if no hardware primary output is found
BZ: 129061

When no hardware primary output is found, there is no audio output
descriptor associated to it. This descriptor shall not be used in
that case as it will lead to a segmentation fault.

Add a LOG_ALWAYS_FATAL_IF to check the hw output descriptor before using it.
It will generate a tombstone and restart the media server if it is a null pointer.

Change-Id: I4cd0328db6b7ccf93b8f56109e936a4d2f867478
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>
Signed-off-by: David.Coutherut <davidx.coutherut@intel.com>
Reviewed-on: http://android.intel.com:8080/123448
Reviewed-by: Di Folco, Neil <neil.di.folco@intel.com>
Reviewed-by: Gonzalve, Sebastien <sebastien.gonzalve@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>