[Audio][HAL]quick release of lock in case of route issue
BZ: 113305

Because an audio route can be unavailable, it is possible that the successive
locks prevent the appropriate routing and in this case, the audio can be played
on speaker instead of earpiece, for example or the remote receives silence.

This patch purpose is to free the lock in order to allow the routing.

Change-Id: I82db96883a1b7c0f122256b5f2e043a486ea8a26
Signed-off-by: Andrei V. FOMITCHEV <andreix.fomitchev@intel.com>
Reviewed-on: http://android.intel.com:8080/128256
Reviewed-by: Gaffie, Francois <francois.gaffie@intel.com>
Reviewed-by: Becker, VincentX <vincentx.becker@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: Quintero, Jorge <jorge.quintero@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>