[AUDIO] Identify correctly the active audio input source
BZ: 132834

After integration of I2dd7ecd84aefdc2c7a0a4ef4c9fb706630b7b69a, the
AudioRouteManager lost the capability to correctly identify and set the
active input source.

For audio input sources, some streams may have a null device to notify
that they lost the ability to be routed. In order to differentiate between
2 audio source inputs and reroute the audio input stream from one source
input to another as expected, this patch adds the logic to be able to
identify which is the active input source.

Change-Id: I239d10bac67a86d6e41f4866f279dd482f656901
Signed-off-by: Stefan BOROS <stefan.petrex.boros@intel.com>
Reviewed-on: http://android.intel.com:8080/128822
Reviewed-by: Fomitchev, AndreiX <andreix.fomitchev@intel.com>
Reviewed-by: Gaffie, Francois <francois.gaffie@intel.com>
Reviewed-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: Di Folco, Neil <neil.di.folco@intel.com>
Reviewed-by: Quintero, Jorge <jorge.quintero@intel.com>
Reviewed-by: Gonzalve, Sebastien <sebastien.gonzalve@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>