[AUDIO] Input stream buffer size rework.
BZ: 118092

Transfer between AudioFlinger and AudioHAL in input are done
using full ring buffer or one period (in communication).
This buffering model has been introduce to support low performance
platform like lexington.

This patch removes this workaround and aligns the behavior of the input
on the ouput (4 periods ring buffer, read transfer done at the
granularity of one period). The only exception to this rule is the
deep buffer configuration.

Change-Id: Iec178eba8aba69a4133783d82ccc61ce6493c768
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/130516
Reviewed-by: Quintero, Jorge <jorge.quintero@intel.com>
Reviewed-by: Di Folco, Neil <neil.di.folco@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>