[Audio] reduce latency by launching a PFW configuration at startup
BZ: 137829

The issue appears during the first media playback after the media server
starts. The cold output latency is bigger than the "Android 4.3
Compatibility Definition" requirement. This SHOULD be below 100 ms.
cf paragraph 4.5 about Audio Latency.

This is partly due to the Audio Parameter Framework that applies all
the parameters during first routing as default values.
It will take almost 200 ms extra latency to do this first
parameters application.

The patch reduces this cold output latency by setting a mute criterion for
output and input so all the default configurations are applied and
the PFW will be ready for the first media routing.

Change-Id: I125b2c97df0ea92e7be5553be50fdcb94e746b6a
Signed-off-by: pafonsoX <philippex.afonso@intel.com>
Reviewed-on: http://android.intel.com:8080/132044
Reviewed-by: Quintero, Jorge <jorge.quintero@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>