[AUDIO] Configure voice mixing capture and playback paths
BZ: 151526

Today, we can ear audio playback in call only on the left channel.
Moreover, the PCM configuration is wrong for audio capture in call.

This patch will fix the audio HAL wrong audio remapping for
audio playback in call.
This patch will fix the PCM configuration for audio capture
in call.

Change-Id: I18244456cfbd2e9dba7a5bb3ea4eb57d2ff15f2a
Signed-off-by: Neil Di Folco <neil.di.folco@intel.com>