[AUDIO] Unset capture route for voice uplink record and mic mute
BZ: 161252

When input source voice uplink is selected to record only uplink
voice during a CSV call and the microphone is muted, the uplink
voice is still recorded. Only silence shall be recorded for the time
the microphone is muted.

Set the media record route to not applicable in those conditions, so
that silence is recorded.

Change-Id: Ia3fd6c059fff0ec579d854ca13c66d11d1ea4214
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>