[AUDIO] Replace INDEX_TO_MASK macro by indexToMask utility method
BZ: 161252

INDEX_TO_MASK is used to convert a bit index into a bitfield mask.
This macro is not 32 bits safe and a check would be needed to check
that the index lies on 32 bits.

This is achieved now thanks to the indexToMask utility method, brought
by the audiocomms-utilities library.
Replacing all occurences of INDEX_TO_MASK usage with indexToMask method.

Change-Id: I83eabd552d90c653046efd0a3392595b69965efd
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>