[AUDIO] No need to access current device in a locked context
BZ: 167974

When reconfiguring the output device at configure stage,
set/get the current device is called in a locked context. This
can lead to a deadlock situation when a write operation is currently
ongoing and is stucked. This may happen on a write during the
disable stage, just before the configure stage.

Remove the call to setCurrentDevices and getCurrentDevices in
a locked context. This is not necessary as a stream route write/read
operation can be locked and prevent these calls and the configure
stage to go on.

Change-Id: If9107c654888e68fb49a10d71a9d58c76ffbf02a
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>