[WORKAROUND][AUDIO] Declare CmeVoice route and InImsCall mode
BZ: 177022

The routemanager legacy is not declaring the CmeVoice route and
InImsCall mode. This cause compatibility issue with the audio parameter
framework used with the XML HAL, and the phone doesn't boot.

To avoid this issue, we just declare the new routes and mode in the routemanager
even if they are not exercized.

Change-Id: Ib5ef795ccef8d521e0a8d7a4f71a114e04fd6adb
Signed-off-by: Michael Soares <michaelx.soares@intel.com>