[BYT FFRD8] AutomaticGainControl/NoiseSuppressor can't be applied when recording
BZ: 178156

The buffer size for audio capture is calculated based on the framecount.
The values of framecount for sampling rates of 11025, 22050 and 44100 are
changed to nearest higher values which are multiples of 16. Buffer sizes are
calculated based on these changed framecounts for above mentioned sampling rates
due to which noise is introduced in each captured buffer and the recorded file
will have noise. This patch will not allow the framecount to be changed so that
correct buffer size is calculated and there wont be noise in the buffer.

Change-Id: Ife0df9c6003db790087c38274672414dd9e0ab80
Category: device enablement
Domain: AUDIO_MEDIA-External
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: charitardha <satya.charitardha.jayanti@intel.com>