[AUDIO] Fix the no audio UL issue for VOIP with BT headset use case
BZ: 181619

When doing a VOIP call with BT headset just after boot, the first route
availability check always return false.
This will return a NULL pointer handle as echo reference handle.
This will cause the AEC audio effect process to return an error.

Now, we take the first output stream without testing whether it's routed
or not as we consider it as a primary output stream.

Change-Id: I646518f23e3d8e7c1b94836b98ce5812ae519d7a
Signed-off-by: li peizhao <peizhao.li@intel.com>